backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.2.0
    # via kombu
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
billiard==4.2.1
    # via celery
celery==5.4.0
    # via -r requirements.in
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   typer
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi[standard]==0.115.2
    # via -r requirements.in
fastapi-cli[standard]==0.0.5
    # via fastapi
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.6
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.27.2
    # via
    #   fastapi
    #   openai
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via fastapi
jiter==0.6.1
    # via openai
kombu==5.4.2
    # via celery
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.1
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
openai==1.51.2
    # via -r requirements.in
packaging==24.1
    # via pytest
pluggy==1.5.0
    # via pytest
prompt-toolkit==3.0.48
    # via click-repl
pydantic==2.9.2
    # via
    #   fastapi
    #   openai
    #   sqlmodel
pydantic-core==2.23.4
    # via pydantic
pygments==2.18.0
    # via rich
pytest==8.3.3
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via celery
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.12
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
rich==13.9.2
    # via typer
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.36
    # via sqlmodel
sqlmodel==0.0.22
    # via -r requirements.in
starlette==0.40.0
    # via fastapi
tqdm==4.66.5
    # via openai
typer==0.12.5
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
tzdata==2024.2
    # via
    #   celery
    #   kombu
uvicorn[standard]==0.32.0
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
watchfiles==0.24.0
    # via uvicorn
wcwidth==0.2.13
    # via prompt-toolkit
websockets==13.1
    # via uvicorn
